requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coverage==7.2.3
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2
django-cors-headers==3.14.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-spectacular==0.26.1
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.4.0
flake8==6.0.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
jsonschema==4.17.3
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
pluggy==1.0.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.15.0
PyJWT==2.6.0
pyrsistent==0.19.3
pytest==7.2.2
pytest-cov==4.0.0
pytest-cover==3.0.0
pytest-coverage==0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.1
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
